[v2] refactor: widen NodeId to u64

+++ crates/solidity-v2/outputs/cargo/ir/src/tests/builder.rs
@@ -139,7 +139,7 @@ contract MyContract {
 
     // Every allocated node is recorded exactly once, so the histogram total
     // must match the number of IDs handed out.
-    assert_eq!(histogram.total() as usize, id_generator.allocated_count());
+    assert_eq!(u64::from(histogram.total()), id_generator.allocated_count());
 
     // Spot-check a few known kinds in the snippet above.
     assert_eq!(1, histogram.count(ir::NodeKind::ContractDefinition));
